extracted container with slider
