+++ src/app/components/common/ContainerWithSlider/index.js
@@ -1,0 +1,1 @@
+export { default } from './ContainerWithSlider';
