+++ src/app/components/userProfile/common/UserNavigation/UserNavigation.jsx
@@ -8,18 +8,9 @@ import { Link } from 'react-router';
 import { TabLink as StyledTabLink, TabLinkIndex as StyledTabLinkIndex } from 'golos-ui/Tab';
 import Icon from 'golos-ui/Icon';
 
-import Container from 'src/app/components/common/Container';
 import { changeProfileLayout } from 'src/app/redux/actions/ui';
+import ContainerWithSlider from '../../../common/ContainerWithSlider/ContainerWithSlider';
 import * as ReactDOM from 'react-dom';
-
-const Wrapper = styled.div`
-    position: relative;
-    min-height: 50px;
-    background-color: #fff;
-    box-shadow: 0 2px 12px 0 rgba(0, 0, 0, 0.15);
-    z-index: 1;
-    overflow-x: hidden;
-`;
 
 const TabLink = styled(StyledTabLink)`
     &.${({ activeClassName }) => activeClassName} {
@@ -82,53 +73,14 @@ const SimpleIcon = styled(Icon)`
     height: 20px;
 `;
 
-const ArrowContainer = styled.div`
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    width: 10px;
-    height: 100%;
-    position: absolute;
-    top: 0;
-    cursor: pointer;
-    z-index: 1;
-`;
-
-const LeftArrowContainer = styled(ArrowContainer)`
-    left: 20px;
-`;
-
-const RightArrowContianer = styled(ArrowContainer)`
-    right: 20px;
-`;
-
-const ArrowIcon = styled(Icon)`
-    width: 8px;
-    height: 14px;
-`;
-ArrowIcon.defaultProps = {
-    name: 'chevron',
-};
-
-const LeftArrow = styled(ArrowIcon)`
-    transform: rotate(-90deg);
-`;
-
-const RightArrow = styled(ArrowIcon)`
-    transform: rotate(90deg);
-`;
-
 class UserNavigation extends PureComponent {
     constructor() {
         super();
         this.state = {
             screenLessThenMainContainer: false,
-            currentOffsetIndex: 0,
         };
-        this._setStyleForIconsNavigation = this._setStyleForIconsNavigation.bind(this);
-        this.showNextIcon = this.showNextIcon.bind(this);
-        this.showPrevIcon = this.showPrevIcon.bind(this);
     }
+
     static propTypes = {
         accountName: PropTypes.string,
         isOwner: PropTypes.bool,
@@ -137,10 +89,6 @@ class UserNavigation extends PureComponent {
         changeProfileLayout: PropTypes.func,
     };
 
-    componentDidUpdate() {
-        this._setStyleForIconsNavigation();
-    }
-
     componentDidMount() {
         this._checkScreenSize();
         window.addEventListener('resize', this._checkScreenSize);
@@ -148,20 +96,6 @@ class UserNavigation extends PureComponent {
 
     componentWillUnmount() {
         window.removeEventListener('resize', this._checkScreenSize);
-    }
-
-    showNextIcon(e) {
-        e.stopPropagation();
-        this.setState({
-            currentOffsetIndex: this.state.currentOffsetIndex + 1,
-        });
-    }
-
-    showPrevIcon(e) {
-        e.stopPropagation();
-        this.setState({
-            currentOffsetIndex: this.state.currentOffsetIndex - 1,
-        });
     }
 
     render() {
@@ -197,34 +131,24 @@ class UserNavigation extends PureComponent {
             },
         ];
         return (
-            <Wrapper className={className} innerRef={ref => (this.wrapper = ref)}>
-                {this.state.currentOffsetIndex > 0 && (
-                    <LeftArrowContainer onClick={this.showPrevIcon}>
-                        <LeftArrow />
-                    </LeftArrowContainer>
-                )}
-                <Container align="center" ref={ref => (this.container = ref)}>
-                    {tabLinks.map(({ value, to }) => (
-                        <TabLinkIndex key={to} to={to}>
-                            {value}
-                        </TabLinkIndex>
-                    ))}
-                    {/*<LinkWithDropdown*/}
-                    {/*closeOnClickOutside*/}
-                    {/*dropdownPosition="bottom"*/}
-                    {/*dropdownContent={<VerticalMenu items={rewardsMenu} />}*/}
-                    {/*>*/}
-                    {/*<TabLink>{tt('g.rewards')}</TabLink>*/}
-                    {/*</LinkWithDropdown>*/}
-                    {this._renderRightIcons()}
-                </Container>
-                <RightArrowContianer
-                    ref={ref => (this.rightArrow = ref)}
-                    onClick={this.showNextIcon}
-                >
-                    <RightArrow />
-                </RightArrowContianer>
-            </Wrapper>
+            <ContainerWithSlider
+                className={className}
+                ref={ref => (this.wrapper = ReactDOM.findDOMNode(ref))}
+            >
+                {tabLinks.map(({ value, to }) => (
+                    <TabLinkIndex key={to} to={to}>
+                        {value}
+                    </TabLinkIndex>
+                ))}
+                {/*<LinkWithDropdown*/}
+                {/*closeOnClickOutside*/}
+                {/*dropdownPosition="bottom"*/}
+                {/*dropdownContent={<VerticalMenu items={rewardsMenu} />}*/}
+                {/*>*/}
+                {/*<TabLink>{tt('g.rewards')}</TabLink>*/}
+                {/*</LinkWithDropdown>*/}
+                {this._renderRightIcons()}
+            </ContainerWithSlider>
         );
     }
 
@@ -266,61 +190,6 @@ class UserNavigation extends PureComponent {
         }
     }
 
-    _setStyleForIconsNavigation() {
-        let container = ReactDOM.findDOMNode(this.container);
-        let rightArrow = ReactDOM.findDOMNode(this.rightArrow);
-        let children = container.children;
-        let currentOffsetIndex = this.state.currentOffsetIndex;
-        let currentOffset = children[currentOffsetIndex];
-
-        const RIGHT_PADDING = 20;
-        const LEFT_PADDING = currentOffsetIndex > 0 ? 10 : 0;
-        const LEFT = children[0].offsetLeft - currentOffset.offsetLeft + LEFT_PADDING;
-
-        rightArrow.style.display = 'none';
-
-        if (container.clientWidth >= container.scrollWidth - LEFT) {
-            this._clearStyleForIconsNavigation(children);
-            return;
-        }
-
-        children[0].style.marginLeft = LEFT + 'px';
-
-        for (let i = 0; i < children.length; i++) {
-            let child = children[i];
-
-            if (i < currentOffsetIndex) {
-                child.style.opacity = 0;
-                continue;
-            }
-
-            if (
-                child.offsetLeft + child.clientWidth - currentOffset.offsetLeft >
-                container.clientWidth - (i < children.length - 1 ? RIGHT_PADDING : 0)
-            ) {
-                child.style.opacity = 0;
-                rightArrow.style.display = 'flex';
-            } else {
-                child.style.opacity = 1;
-            }
-        }
-    }
-
-    _clearStyleForIconsNavigation(children) {
-        for (let i = 0; i < children.length; i++) {
-            let child = children[i];
-            if (i === 0) {
-                child.style.marginLeft = 0;
-            }
-            child.style.opacity = 1;
-        }
-        if (this.state.currentOffsetIndex !== 0) {
-            this.setState({
-                currentOffsetIndex: 0,
-            });
-        }
-    }
-
     _onGridClick = () => {
         this.props.changeProfileLayout('grid');
     };
@@ -346,8 +215,6 @@ class UserNavigation extends PureComponent {
         ) {
             this.setState({ screenLessThenMainContainer: false });
         }
-
-        this._setStyleForIconsNavigation();
     };
 }
